backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.21
aiosqlite==0.19.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0
openai==1.3.5
async-timeout==4.0.3
loguru==0.7.2